requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
et-xmlfile==1.1.0
gunicorn==22.0.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
